new load buttons

+++ content.js
@@ -1,10 +1,12 @@
-ready();
+let load;
+ready("load_newer_button");
+
 let el_button = document.getElementById("load-more-button");
 if (el_button) {
-    console.log('el_button');
+    //console.log('el_button');
     document.getElementById("load-more-button").addEventListener("click", function() {
-        console.log('el_button-ready');
-        ready();
+        //console.log('el_button-ready');
+        ready("load-more-button");
     })
 }
 
@@ -14,11 +16,11 @@ let check_myStopFunction;
 check_myStopFunction = setInterval(function(){
     let el = document.querySelector('.load-newer-button');
     if (el) {
-        console.log('ready-true');
+        //console.log('ready-true');
         load_newer_button ();
         myStopFunction();
     }
-    console.log('setInterval');
+    //console.log('setInterval');
 }, 6000);
 
 
@@ -29,37 +31,42 @@ function myStopFunction() {
 
 function load_newer_button () {
     document.querySelector(".load-newer-button").addEventListener("click", function () {
-        ready();
-        console.log('click-load-newer-button');
+        ready("load_newer_button");
+        //console.log('click-load-newer-button');
     });
 }
 
-function ready () {
-    console.log('ready-start');
+function ready (load) {
+    console.log(load);
 
     setTimeout(function() {
-        console.log('ready-onload');
+        //console.log('ready-onload');
         let last_save = localStorage.getItem('last_link');
-        console.log('last_save-' + last_save);
+        //console.log('last_save-' + last_save);
         let last = document.querySelector("section.job-tile");
         let last_link = last.querySelector("h4 a");
         last_link = last_link.href;
         if (!last_link) {
             console.log("last_link " + last_link);
-            ready();
+            ready("load_newer_button");
         }
 
         let all = document.querySelectorAll("section.job-tile");
-        localStorage.setItem('last_link', last_link);
+        let ALREADYREAD = document.getElementById("ALREADYREAD");
+        console.log(ALREADYREAD);
+        if(load=="load_newer_button" && !ALREADYREAD) {
+            localStorage.setItem('last_link', last_link);
+            console.log('set localStorage');
+        }
         let arrayLength = all.length;
-        console.log('arrayLength-' + arrayLength);
+        //console.log('arrayLength-' + arrayLength);
         for (var i = 0; i < arrayLength; i++) {
-            console.log('i-' + i);
+            //console.log('i-' + i);
             let arr_i_href = all[i].querySelector("h4 a").href
             if (last_save == arr_i_href) {
-                let ALREADYREAD = document.getElementById("ALREADYREAD");
+                ALREADYREAD = document.getElementById("ALREADYREAD");
                 if (!ALREADYREAD) {
-                    console.log('last_save == arr_i_href-');
+                    //console.log('last_save == arr_i_href-');
                     let set = all[i].querySelector(".job-title")
                     set.outerHTML = "<h2 id='ALREADYREAD' style=' width: 100%;  text-align: center; border-bottom: 1px solid #1d4354;  line-height: 0.1em; margin: -20px 0 50px; '>" +
                         "<span style='color: #14a800;background:#fff; padding:0 10px;'>ALREADY READ</span></h2> " + set.outerHTML;
